LPD-37537 Separate to new test file to prepare for additional tests and refactoring
