+++ modules/apps/site-navigation/site-navigation-test/src/testIntegration/java/com/liferay/site/navigation/menu/web/internal/exportimport/test/SiteNavigationMenuExportImportTest.java
@@ -8,14 +8,11 @@ package com.liferay.site.navigation.menu.web.internal.exportimport.test;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationParameterMapFactoryUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
-import com.liferay.exportimport.kernel.staging.MergeLayoutPrototypesThreadLocal;
 import com.liferay.exportimport.kernel.staging.StagingUtil;
 import com.liferay.layout.test.util.LayoutTestUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutSet;
-import com.liferay.portal.kernel.model.LayoutSetPrototype;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.service.PortletPreferencesLocalService;
@@ -23,10 +20,8 @@ import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.PortletKeys;
-import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portlet.display.template.test.util.BaseExportImportTestCase;
 import com.liferay.site.navigation.constants.SiteNavigationMenuPortletKeys;
@@ -151,69 +146,6 @@ public class SiteNavigationMenuExportImportTest
 			"1",
 			portletPreferences.getValue(
 				"siteNavigationMenuType", StringPool.BLANK));
-	}
-
-	@Test
-	public void testSiteTemplatePropagationWhenDuplicateSiteNavigationMenusExist()
-		throws Exception {
-
-		Group group = GroupTestUtil.addGroup();
-
-		LayoutSetPrototype layoutSetPrototype =
-			LayoutTestUtil.addLayoutSetPrototype(RandomTestUtil.randomString());
-
-		Layout prototypeLayout = LayoutTestUtil.addTypePortletLayout(
-			layoutSetPrototype.getGroup(), true);
-
-		MergeLayoutPrototypesThreadLocal.clearMergeComplete();
-
-		_sites.updateLayoutSetPrototypesLinks(
-			group, layoutSetPrototype.getLayoutSetPrototypeId(), 0, true, true);
-
-		Layout layout = _layoutLocalService.getFriendlyURLLayout(
-			group.getGroupId(), false, prototypeLayout.getFriendlyURL());
-
-		layout.setLayoutPrototypeUuid(prototypeLayout.getUuid());
-		layout.setLayoutPrototypeLinkEnabled(true);
-
-		_layoutLocalService.updateLayout(layout);
-
-		String name = RandomTestUtil.randomString();
-
-		SiteNavigationMenuTestUtil.addSiteNavigationMenu(group, name);
-
-		LayoutTestUtil.addPortletToLayout(
-			prototypeLayout, SiteNavigationMenuPortletKeys.SITE_NAVIGATION_MENU,
-			HashMapBuilder.put(
-				"siteNavigationMenuExternalReferenceCode",
-				() -> {
-					SiteNavigationMenu siteNavigationMenu =
-						SiteNavigationMenuTestUtil.addSiteNavigationMenu(
-							layoutSetPrototype.getGroup(), name);
-
-					return new String[] {
-						siteNavigationMenu.getExternalReferenceCode()
-					};
-				}
-			).build());
-
-		MergeLayoutPrototypesThreadLocal.clearMergeComplete();
-
-		MergeLayoutPrototypesThreadLocal.setSkipMerge(false);
-
-		_sites.mergeLayoutSetPrototypeLayouts(
-			group, group.getPublicLayoutSet());
-
-		LayoutSet layoutSet = layoutSetPrototype.getLayoutSet();
-
-		UnicodeProperties layoutSetPrototypeSettingsUnicodeProperties =
-			layoutSet.getSettingsProperties();
-
-		Assert.assertEquals(
-			0,
-			GetterUtil.getInteger(
-				layoutSetPrototypeSettingsUnicodeProperties.getProperty(
-					Sites.MERGE_FAIL_COUNT)));
 	}
 
 	private void _publishLayouts() throws Exception {
